Return car service
Implement an endpoint ‘return-car’ where a customer-id, car-id is passed as parameters. Car’s status (e.g., ok or damaged) during the return will also be sent as a parameter. The system must check that the customer with customer-id has rented the car. The car’s status is changed from ‘booked’ to ‘available’ or ‘damaged’.